moved viewmanager
