+++ spoticli/Components/Main/MainMenu.py
@@ -1,4 +1,4 @@
-from ViewManager import viewManager
+from Components.Main.ViewManager import viewManager
 from Components.Templates.Menu import Menu, MenuItem
 from Components.Main.PlaylistMenu import PlaylistMenu
