+++ spoticli/Components/Main/SongItem.py
@@ -1,4 +1,4 @@
-from ViewManager import viewManager
+from Components.Main.ViewManager import viewManager
 from Components.Templates.Component import Component
 from Components.Templates.Menu import MenuItem
 from Components.Main.Player import player
