+++ spoticli/Components/Main/TitleBar.py
@@ -1,4 +1,4 @@
-from ViewManager import viewManager
+from Components.Main.ViewManager import viewManager
 from Components.Templates.Component import Component
