+++ spoticli/Components/Main/ViewManager.py
