+++ spoticli/Components/Templates/Menu.py
@@ -1,5 +1,5 @@
 import math
-from ViewManager import viewManager
+from Components.Main.ViewManager import viewManager
 from Components.Templates.Component import Component
 from Components.Main.Player import player
